lifecompass-ai-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.3
python-dotenv==1.0.0
pydantic>=2.0.0
python-multipart==0.0.6
requests==2.31.0

# AI Provider Dependencies (Optional - install based on your choice)
# Google AI
google-generativeai==0.3.2

# OpenAI (Optional)
# openai==1.3.0

# Anthropic (Optional)  
# anthropic==0.7.0

# Hugging Face (Optional)
# transformers==4.35.0
